Moved ShowPlaylist to application Package. Finished library option. Need to expand playlist
