+++ src/main/java/ie/atu/jdbc/application/ShowPlaylist.java
@@ -1,22 +1,30 @@
-package ie.atu.jdbc.pool;
+package ie.atu.jdbc.application;
 
 import ie.atu.jdbc.pool.DatabaseUtils;
 import java.sql.*;
 import java.util.ArrayList;
 public class ShowPlaylist {
-    public static void main(String[] args) {
+    public static void main(String[] args){
+        String username = "";
+        ShowPlaylist(username);
+    }
+    public static void ShowPlaylist(String username) {
         String selectSQL = "SELECT user_playlist.playlist_name, users.name  " +
                 "FROM user_playlist " +
                 "JOIN users ON user_playlist.user_id = users.user_id " +
-                "WHERE users.user_id = 20";
+                "WHERE users.username= ?";
+
         ArrayList<String> playlists = new ArrayList<>();
+
         try (Connection connection = DatabaseUtils.getConnection();
-             Statement statement = connection.createStatement();
-             ResultSet resultSet = statement.executeQuery(selectSQL)) {
-            while (resultSet.next()) {
-                String playlistName = resultSet.getString("playlist_name");
-                String userName = resultSet.getString("name");
-                playlists.add(playlistName + " - "+ userName);
+             PreparedStatement statement = connection.prepareStatement(selectSQL)){
+            statement.setString(1,username);
+
+            ResultSet resultSetPlaylist = statement.executeQuery();
+            while (resultSetPlaylist.next()) {
+                String playlistName = resultSetPlaylist.getString("playlist_name");
+                String userName = resultSetPlaylist.getString("name");
+                playlists.add(playlistName + " - "+ username);
             }
             if(playlists.size()==0){
                 System.out.println("No playlists\n");
